+++ src/coremods/COREMOD_iofits/images2cube.h
@@ -7,5 +7,5 @@ errno_t CLIADDCMD_COREMOD_iofits__images2cube();
 
 errno_t images_to_cube(
     const char *restrict img_name,
-    long nbframes,
+    uint32_t nbframes,
     const char *restrict cube_name);
